BUG FIX: convertCel() on a CCG/v1 CEL file could give errors [#11]
"Error in sprintf("GridCorner%s=%d %d\n" ... invalid format '%d' ...)".
Added package test for convertCel(), but in this particular case
it would not have cought it because it only happened for chip types
of particular dimensions, where the dimension stored as a 4-byte float
would be read into R as a "near-integer" double, e.g.
x - 460 == 3.051758e-05.